HAI-183: refaktoroitu geometrian käsittely omaan Controlleriin, jotta ei sotkisi kehitystyötä HankeControllerin kanssa
