+++ services/hanke-service/src/integrationTest/kotlin/fi/hel/haitaton/hanke/HankeControllerITests.kt
@@ -1,6 +1,6 @@
 package fi.hel.haitaton.hanke
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.databind.ObjectMapper
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration
@@ -13,8 +13,6 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-import java.nio.file.Files
-import java.nio.file.Paths
 
 /**
  * Testing the Hnake Controller through a full REST request.
@@ -26,8 +24,6 @@ import java.nio.file.Paths
 @AutoConfigureMockMvc
 @EnableAutoConfiguration
 class HankeControllerITests(@Autowired val mockMvc: MockMvc) {
-
-    private val objectMapper = jacksonObjectMapper()
 
     @Test
     fun `When hankeId not given then Bad Request`() {
@@ -54,6 +50,7 @@ class HankeControllerITests(@Autowired val mockMvc: MockMvc) {
         //make Hanke for request (timestamps are Strings for now)
         val hankeToBeAdded = RestDataHanke("kissa", hankeName, "20201212", "20201212", "Risto", 1)
 
+        val objectMapper = ObjectMapper()
         val hankeJSON = objectMapper.writeValueAsString(hankeToBeAdded)
 
         mockMvc.perform(post("/hankkeet").contentType(MediaType.APPLICATION_JSON).content(hankeJSON)
@@ -63,39 +60,6 @@ class HankeControllerITests(@Autowired val mockMvc: MockMvc) {
                 .andExpect(MockMvcResultMatchers.jsonPath("$.name").value(hankeName))
     }
 
-    @Test
-    fun `create Geometria OK`() {
-        val content = Files.readString(Paths.get("src/test/resources/featureCollection.json"))
-
-        mockMvc.perform(post("/hankkeet/1234567/geometriat")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(content)
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNoContent)
-                .andExpect(content().string(""))
-    }
-
-    @Test
-    fun `create Geometria with invalid HankeID`() {
-        val content = Files.readString(Paths.get("src/test/resources/featureCollection.json"))
-
-        mockMvc.perform(post("/hankkeet/INVALID/geometriat")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(content)
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound)
-                .andExpect(MockMvcResultMatchers.jsonPath("$.errorCode").value("HAI1001"))
-    }
-
-    @Test
-    fun `create Geometria without Geometria`() {
-        mockMvc.perform(post("/hankkeet/INVALID/geometriat")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content("")
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isBadRequest)
-                .andExpect(MockMvcResultMatchers.jsonPath("$.errorCode").value("HAI1011"))
-    }
 }
