+++ services/hanke-service/src/integrationTest/kotlin/fi/hel/haitaton/hanke/hello/HelloApiControllerITests.kt
@@ -3,7 +3,10 @@ package fi.hel.haitaton.hanke.hello
 import org.hamcrest.Matchers.stringContainsInOrder
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
+import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
@@ -16,7 +19,9 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPat
  *
  * This class should test only the weblayer (both HTTP server and context to be auto-mocked).
  */
-@WebMvcTest
+@SpringBootTest
+@AutoConfigureMockMvc
+@EnableAutoConfiguration
 class HelloApiControllerITests(@Autowired val mockMvc : MockMvc) {
 
     @Test
