+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/Constants.kt
@@ -2,4 +2,4 @@ package fi.hel.haitaton.hanke
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 
-internal val objectMapper = jacksonObjectMapper()
+val objectMapper = jacksonObjectMapper()
